feat: add BaseBallGameNumber Class
모든 원시값과 문자열을 포장한다. 해당 클래스에서 유효성 검사 하도록 수정
